scripts paper
