--- scripts_adaptation/run_baselines.sh
@@ -1,3 +1,0 @@
-python train.py task=TrackRand algo=ppo headless=true total_frames=150000000
-python train.py task=TrackRand algo=ppo_gru headless=true total_frames=150000000
-python train.py task=TrackRand algo=ppo_adapt headless=true total_frames=150000000
