+++ scripts_paper/experiments/drone_model.yaml
